requirements.txt
fastapi
uvicorn[standard]
python-multipart
pdfplumber
python-docx
scikit-learn
weasyprint
python-dotenv
